[frontend] Show run info in run comparison
