+++ frontend/src/components/rundetail/RunDetail.vue
@@ -27,37 +27,7 @@
     </v-row>
     <v-row no-gutters class="mt-3">
       <v-col>
-        <v-card>
-          <v-card-title>
-            <v-toolbar dark color="primary">
-              Run information
-            </v-toolbar>
-          </v-card-title>
-          <v-card-text class="py-0">
-            <v-container fluid class="ma-0 pa-0">
-              <v-row align="center" justify="space-around">
-                <v-col
-                  :lg="item.alwaysAuto ? 'auto' : '2'"
-                  md="auto"
-                  sm="auto"
-                  class="pt-0"
-                  v-for="item in runInfoItems"
-                  :key="item.header"
-                >
-                  <v-card outlined>
-                    <v-card-title class="pb-1">
-                      <v-icon left dense>{{ item.icon }}</v-icon>
-                      {{ item.header }}
-                    </v-card-title>
-                    <v-card-text>
-                      <span :class="item.bodyClass">{{ item.body }}</span>
-                    </v-card-text>
-                  </v-card>
-                </v-col>
-              </v-row>
-            </v-container>
-          </v-card-text>
-        </v-card>
+        <run-info :run="run"></run-info>
       </v-col>
     </v-row>
   </v-container>
@@ -76,18 +46,12 @@ import {
   RunWithDifferences
 } from '@/store/types'
 import { Prop } from 'vue-property-decorator'
-import { formatDate, formatDurationHuman } from '@/util/TimeUtil'
 import MeasurementsDisplay from '@/components/rundetail/MeasurementsDisplay.vue'
-import {
-  mdiFlash,
-  mdiCameraTimer,
-  mdiAlarmCheck,
-  mdiClockFast,
-  mdiRobot
-} from '@mdi/js'
+import RunInfo from '@/components/rundetail/RunInfo.vue'
 
 @Component({
   components: {
+    'run-info': RunInfo,
     'measurements-display': MeasurementsDisplay
   }
 })
@@ -95,50 +59,8 @@ export default class RunDetail extends Vue {
   @Prop()
   private runWithDifferences!: RunWithDifferences
 
-  private formatDate(date: Date) {
-    return formatDate(date)
-  }
-
-  private formatDuration(start: Date, end: Date) {
-    return formatDurationHuman(start, end)
-  }
-
   private get run(): Run {
     return this.runWithDifferences.run
-  }
-
-  private get runInfoItems() {
-    return [
-      {
-        header: 'Trigger',
-        icon: this.iconTrigger,
-        body: `${this.run.source.type.toLocaleLowerCase()} by ${
-          this.run.author
-        }`
-      },
-      {
-        header: 'Started',
-        icon: this.iconStarted,
-        body: this.formatDate(this.run.startTime)
-      },
-      {
-        header: 'Finished',
-        icon: this.iconFinished,
-        body: this.formatDate(this.run.stopTime)
-      },
-      {
-        header: 'Duration',
-        icon: this.iconDuration,
-        body: this.formatDuration(this.run.startTime, this.run.stopTime)
-      },
-      {
-        header: this.run.runnerName,
-        icon: this.iconRunner,
-        body: this.run.runnerInfo,
-        bodyClass: 'worker-description',
-        alwaysAuto: true
-      }
-    ]
   }
 
   private get runColor() {
@@ -180,18 +102,10 @@ export default class RunDetail extends Vue {
     }
     return undefined
   }
-
-  // ICONS
-  private iconTrigger = mdiFlash
-  private iconStarted = mdiClockFast
-  private iconFinished = mdiAlarmCheck
-  private iconDuration = mdiCameraTimer
-  private iconRunner = mdiRobot
 }
 </script>
 
 <style scoped>
-.worker-description,
 .error-text {
   white-space: pre-wrap;
   font-family: monospace;
